Tile rendering added
Making a 16x12 map of numbers in a text file, you can now create maps. The tileManager reads the txt file, and gets the data as a strings, splits it and the parses it to a integer. This can now be used together with the tile array (Stores the images for the tiles), to draw specifik tiles, based on the content of the txt file

Added
Tile.java
TileManager.java